BREAKING CHANGE:
Regrettably, I had to do a breaking change, since the HTTP 307 does not
work like I thought. HTTP clients (like Chrome) send the full body of
the PUT request, despite of receiving a HTTP 307 with a location header
for a redirect. For that reason the segment's content is effectively
send twice. To prevent this double upload, the upload-manager now sends
an HTTP GET request to receive an upload-url, before its send a PUT
request to this URL with the segment's content.